게시판 댓글 기능 추가

+++ pages/api/post/delete.js
@@ -1,13 +1,21 @@
 import { connectDB } from "../../../src/app/util/database"
 import { ObjectId } from "mongodb";
+import { getServerSession } from "next-auth"
+import { authOptions } from "../auth/[...nextauth]"
 
 export default async function handler(요청, 응답) {
   if (요청.method == 'POST'){
-     
+    let session =await getServerSession(요청,응답,authOptions)
 
-      let db = (await connectDB).db('forum') 
-      let result = await db.collection('post').deleteOne({_id : new ObjectId(요청.body)});
-      
-      응답.status(200).redirect('/list')
+    let db = (await connectDB).db('forum') 
+    let type = await db.collection('post').findOne({_id : new ObjectId(요청.body)});
+
+    if(type.author == session.user.email){
+      let result = await db.collection('post').deleteOne({_id : new ObjectId(요청.body)})
+      return 응답.status(200).json('삭제완료')
+    }else{
+      return 응답.status(500).json('현재 유저와 불일치')
+    }
+
   }
 }
